Move the plugin to the product directory
